+++ src/Data/Profunctor/Grammar.hs
@@ -28,6 +28,7 @@ import Control.Monad.State
 import Control.Lens
 import Control.Lens.Extras
 import Control.Lens.Grammar.BackusNaur
+import Control.Lens.Grammar.Boole
 import Control.Lens.Grammar.Kleene
 import Control.Lens.Grammar.Symbol
 import Control.Lens.Grammar.Token
