+++ test/Spec.hs
@@ -4,6 +4,7 @@ import Data.Char
 import Data.Foldable hiding (toList)
 import Control.Lens.Grammar
 import Control.Lens.Grammar.BackusNaur
+import Control.Lens.Grammar.Boole
 import Control.Lens.Grammar.Kleene
 import Control.Lens.Grammar.Symbol
 import Control.Lens.Grammar.Token
